implemented 32 number input to be possible

+++ include/special_opts.h
@@ -5,7 +5,7 @@
 #ifndef SPECIAL_OPTS_H
 #define SPECIAL_OPTS_H
 
-int8_t *extract_comment_metadata(const char *prgrm_path, uint8_t *len) ;
+uint32_t *extract_comment_metadata(const char *prgrm_path, uint8_t *len) ;
 
 void create_out_and_err_file() ;
 void adjust_print(bool is_stdout, const char *format, const char *format_no_newline, ...) ;
